I edit the function. There will be 2.0 version
Though this function (V1.0) works very well, it has some problem when I use it to finish my homework. That is when cascade its output to other function, something bad will happen, due to the coordinates it yields is time reversal. It is not convenient. So I edited this function to make it better.