Version2.0
2018.11.20
When I use V1.0, I found some problem when I cascade its output to other function. That is, though the figure plotted by the function is correct, the time matrix of the output is in reverse.  It is not convenient to use it. So I corrected it and made the function smaller 10KB to 7KB by deleting about 60 lines and optimizing some variations.